FRR: Add manifests for speaker running alongside FRR and the reloader
Adding manifests that integrate FRR and the reloader with the
current speaker deployment.

The manifests were created using the current manifests, adding a sidecar
FRR and reloader containers. They also include the ConfigMap for the initial
FRR files and also an initContainer to set everything up.
Furthermore, shareProcessNamespace: true was added to allow the speaker container
to signal the reloader container, the speaker and reloader share a volume where
it would place the rendered configuration file, while the reloader and FRR share
the FRR configuration and sockets volumes so the reloader can communicate with
the FRR container via vtysh and update its configuration.

These serve as a good starting point for the actual FRR integration but
are currently unusable and considered WIP.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>