FRR: Don't listen bgp port and set connect timer for neighbors
The native bgp implementation does not listen on a bgp port,
meaning it is always the side initiating the BGP connection.
We should mimic this behavior in FRR mode as it preserves some
of the peer's configuration by using:
https://docs.frrouting.org/en/latest/bgp.html#cmdoption-bgpd-p.
For example peer-port: without this change there is no guarantee
that it is actually used for the connection - the other side might
be initiating it hence the peer's port will be random from the speaker's side (TCP).

Overriding the default connect timer for neighbors (120s) with 10s
should be fair enough.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>