Tasks: consume a local controller-gen version
Bumping the manifests relies on the version that is installed on the
local machine. This means, calling it on different machines might lead
to different results, having inconsistently generated files.
Here we install the binary to a local path with a pinned version and we
consume it from there.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>
(cherry picked from commit 117d74941bdbbc48e1a003c82f6ad1aeebf3406f)