BGP: add a SetEventCallback function to the session manager
The real reason why we need this, is to be able to inject a listener
that listens to generated frrk8s changes. It has to be a listener
because it has to added AFTER the creation of the manager, that in turn
is passed to the various controllers.

In order to keep the event generic, we pass an empty interface to the
callback, so the consumer can assert the type and handle the event
accordingly to its type.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>