E2E: add an e2e for receiving prefixes alongside MetalLB
We adjust the test that leverages the bgpextras configmap to also
test receiving prefixes using frr-k8s.
This changes the existing test a bit, such as creating only the peers in the beggining
and validating that a simple announcment works at the end, which adds the benefit of
testing a regular MetalLB workload alongside the configmap/frr-k8s.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>