Controller: Align with pointered hold/keepalive timers
We align the code to use the pointer update, fixing the
frr template to specify the timers only when explicitly stated.
In addition, the native code is hardcoded to put a default as before
in case the timers are not specified.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>
(cherry picked from commit bf4c8d5f707bbebd513f4220ed003fe7213cf726)