E2E: fix preferDualStack flake
The test re-used the same pool name twice, effectively making it racey.
Now that the pool status is implemented we can first verify the config
was loaded to make sure the service does not reuse its existing pool.
In addition, removed the first pool being pinned to the namespace to
not have it with a higher priority.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>